requirements.txt
google-generativeai
ddgs
python-dotenv